+++ src/data/script.ts
@@ -1,5 +1,46 @@
 import { HomeAssistant } from "../types";
 import { computeObjectId } from "../common/entity/compute_object_id";
+import { Condition } from "./automation";
+
+export interface EventAction {
+  event: string;
+  event_data?: { [key: string]: any };
+  event_data_template?: { [key: string]: any };
+}
+
+export interface ServiceAction {
+  service: string;
+  entity_id?: string;
+  data?: { [key: string]: any };
+}
+
+export interface DeviceAction {
+  device_id: string;
+  domain: string;
+  entity_id: string;
+}
+
+export interface DelayAction {
+  delay: number;
+}
+
+export interface SceneAction {
+  scene: string;
+}
+
+export interface WaitAction {
+  wait_template: string;
+  timeout?: number;
+}
+
+export type Action =
+  | EventAction
+  | DeviceAction
+  | ServiceAction
+  | Condition
+  | DelayAction
+  | SceneAction
+  | WaitAction;
 
 export const triggerScript = (
   hass: HomeAssistant,
