+++ src/panels/config/automation/action/ha-automation-action.ts
@@ -9,7 +9,7 @@ import {
 } from "lit-element";
 import { fireEvent } from "../../../../common/dom/fire_event";
 import "../../../../components/ha-card";
-import { Action } from "../../../../data/automation";
+import { Action } from "../../../../data/script";
 import { HomeAssistant } from "../../../../types";
 import "./ha-automation-action-row";
