+++ src/panels/config/automation/action/types/ha-automation-action-service.ts
@@ -12,7 +12,7 @@ import { computeDomain } from "../../../../../common/entity/compute_domain";
 import { computeObjectId } from "../../../../../common/entity/compute_object_id";
 import { PolymerChangedEvent } from "../../../../../polymer-types";
 import { fireEvent } from "../../../../../common/dom/fire_event";
-import { ServiceAction } from "../../../../../data/automation";
+import { ServiceAction } from "../../../../../data/script";
 
 @customElement("ha-automation-action-service")
 export class HaServiceAction extends LitElement implements ActionElement {
