+++ src/panels/config/automation/condition/types/ha-automation-condition-device.ts
@@ -28,9 +28,8 @@ export class HaDeviceCondition extends LitElement {
   }
 
   protected render() {
-    if (this._deviceId === undefined) {
-      this._deviceId = this.condition.device_id;
-    }
+    const deviceId = this._deviceId || this.condition.device_id;
+
     const extraFieldsData =
       this._capabilities && this._capabilities.extra_fields
         ? this._capabilities.extra_fields.map((item) => {
@@ -40,14 +39,14 @@ export class HaDeviceCondition extends LitElement {
 
     return html`
       <ha-device-picker
-        .value=${this._deviceId}
+        .value=${deviceId}
         @value-changed=${this._devicePicked}
         .hass=${this.hass}
         label="Device"
       ></ha-device-picker>
       <ha-device-condition-picker
         .value=${this.condition}
-        .deviceId=${this._deviceId}
+        .deviceId=${deviceId}
         @value-changed=${this._deviceConditionPicked}
         .hass=${this.hass}
         label="Condition"
